requirements.txt
fastapi
uvicorn
request
